reorganize test files
