+++ servers/list-api/src/test/graphql/mutations/pocketSave/save-favorite.integration.ts
@@ -1,13 +1,13 @@
-import { readClient, writeClient } from '../../../database/client';
-import { ContextManager } from '../../../server/context';
-import { startServer } from '../../../server/apollo';
+import { readClient, writeClient } from '../../../../database/client';
+import { ContextManager } from '../../../../server/context';
+import { startServer } from '../../../../server/apollo';
 import { Application } from 'express';
 import { ApolloServer } from '@apollo/server';
 import { gql } from 'graphql-tag';
 import { print } from 'graphql';
 import request from 'supertest';
 
-describe('saveUnFavorite mutation', function () {
+describe('saveFavorite mutation', function () {
   const writeDb = writeClient();
   const readDb = readClient();
   const headers = { userid: '1' };
@@ -17,9 +17,9 @@ describe('saveUnFavorite mutation', function () {
   let server: ApolloServer<ContextManager>;
   let url: string;
 
-  const SAVE_UNFAVORITE = gql`
-    mutation saveUnFavorite($id: [ID!]!, $timestamp: ISOString!) {
-      saveUnFavorite(id: $id, timestamp: $timestamp) {
+  const SAVE_FAVORITE = gql`
+    mutation saveFavorite($id: [ID!]!, $timestamp: ISOString!) {
+      saveFavorite(id: $id, timestamp: $timestamp) {
         save {
           id
           favorite
@@ -40,9 +40,10 @@ describe('saveUnFavorite mutation', function () {
   beforeEach(async () => {
     await writeDb('list').truncate();
     const inputData = [
-      { item_id: 0, favorite: 1 },
-      { item_id: 1, favorite: 1 },
-      { item_id: 2, favorite: 0 },
+      { item_id: 0, favorite: 0 },
+      { item_id: 1, favorite: 0 },
+      // One that's already favorited
+      { item_id: 2, favorite: 1 },
     ].map((row) => {
       return {
         ...row,
@@ -53,8 +54,8 @@ describe('saveUnFavorite mutation', function () {
         title: `title ${row.item_id}`,
         time_added: date,
         time_updated: date1,
-        time_read: date,
-        time_favorited: row.favorite === 1 ? date : '0000-00-00 00:00:00',
+        time_read: row.favorite === 1 ? date : '0000-00-00 00:00:00',
+        time_favorited: date,
         api_id: 'apiid',
         api_id_updated: 'apiid',
       };
@@ -75,7 +76,7 @@ describe('saveUnFavorite mutation', function () {
 
   afterEach(() => jest.clearAllMocks());
 
-  it('should unfavorite one save', async () => {
+  it('should favorite one save', async () => {
     const testTimestamp = '2023-10-05T14:48:00.000Z';
     const variables = {
       id: ['1'],
@@ -85,21 +86,21 @@ describe('saveUnFavorite mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(SAVE_UNFAVORITE), variables });
+      .send({ query: print(SAVE_FAVORITE), variables });
 
     expect(res).not.toBeUndefined();
-    expect(res.body.data.saveUnFavorite.save).toBeArrayOfSize(1);
-    expect(res.body.data.saveUnFavorite.errors).toBeArrayOfSize(0);
-    const actual = res.body.data.saveUnFavorite.save[0];
+    expect(res.body.data.saveFavorite.save).toBeArrayOfSize(1);
+    expect(res.body.data.saveFavorite.errors).toBeArrayOfSize(0);
+    const actual = res.body.data.saveFavorite.save[0];
     expect(actual).toStrictEqual({
       id: '1',
-      favorite: false,
-      favoritedAt: null,
+      favorite: true,
+      favoritedAt: testTimestamp,
       updatedAt: testTimestamp,
     });
   });
 
-  it('should unfavorite multiple saves', async () => {
+  it('should favorite multiple saves', async () => {
     const testTimestamp = '2023-10-05T14:48:00.000Z';
     const variables = {
       id: ['0', '1'],
@@ -109,26 +110,26 @@ describe('saveUnFavorite mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(SAVE_UNFAVORITE), variables });
+      .send({ query: print(SAVE_FAVORITE), variables });
 
     const expected = {
       save: [
         {
           id: '0',
-          favorite: false,
-          favoritedAt: null,
+          favorite: true,
+          favoritedAt: testTimestamp,
           updatedAt: testTimestamp,
         },
         {
           id: '1',
-          favorite: false,
-          favoritedAt: null,
+          favorite: true,
+          favoritedAt: testTimestamp,
           updatedAt: testTimestamp,
         },
       ],
       errors: [],
     };
-    const data = res.body.data.saveUnFavorite;
+    const data = res.body.data.saveFavorite;
     expect(data.save).toIncludeSameMembers(expected.save);
     expect(data.errors).toBeArrayOfSize(0);
   });
@@ -143,20 +144,20 @@ describe('saveUnFavorite mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(SAVE_UNFAVORITE), variables });
+      .send({ query: print(SAVE_FAVORITE), variables });
 
     expect(res).not.toBeUndefined();
-    expect(res.body.data.saveUnFavorite.save).toBeArrayOfSize(0);
-    const errors = res.body.data.saveUnFavorite.errors;
+    expect(res.body.data.saveFavorite.save).toBeArrayOfSize(0);
+    const errors = res.body.data.saveFavorite.errors;
     expect(errors).toBeArrayOfSize(1);
     expect(errors[0]).toStrictEqual({
       __typename: 'NotFound',
       message: 'Entity identified by key=id, value=123123 was not found.',
-      path: 'saveUnFavorite',
+      path: 'saveFavorite',
     });
   });
 
-  it('should not fail if trying to unfavorite a save that is already unfavorited (no-op)', async () => {
+  it('should not fail if trying to favorite a save that is already favorited (no-op)', async () => {
     const testTimestamp = '2023-10-05T14:48:00.000Z';
     const variables = {
       id: ['2'],
@@ -166,16 +167,16 @@ describe('saveUnFavorite mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(SAVE_UNFAVORITE), variables });
+      .send({ query: print(SAVE_FAVORITE), variables });
 
-    const data = res.body.data.saveUnFavorite.save;
+    const data = res.body.data.saveFavorite.save;
     expect(data).toMatchObject([
       {
-        favorite: false,
-        favoritedAt: null,
+        favorite: true,
+        favoritedAt: date.toISOString(),
       },
     ]);
-    expect(res.body.data.saveUnFavorite.errors).toBeArrayOfSize(0);
+    expect(res.body.data.saveFavorite.errors).toBeArrayOfSize(0);
   });
   //todo: event emission
   //todo: constraint validation
