+++ servers/list-api/src/test/graphql/mutations/pocketSave/save-unfavorite.integration.ts
@@ -1,13 +1,13 @@
-import { readClient, writeClient } from '../../../database/client';
-import { ContextManager } from '../../../server/context';
-import { startServer } from '../../../server/apollo';
+import { readClient, writeClient } from '../../../../database/client';
+import { ContextManager } from '../../../../server/context';
+import { startServer } from '../../../../server/apollo';
 import { Application } from 'express';
 import { ApolloServer } from '@apollo/server';
 import { gql } from 'graphql-tag';
 import { print } from 'graphql';
 import request from 'supertest';
 
-describe('saveUnArchive mutation', function () {
+describe('saveUnFavorite mutation', function () {
   const writeDb = writeClient();
   const readDb = readClient();
   const headers = { userid: '1' };
@@ -17,13 +17,13 @@ describe('saveUnArchive mutation', function () {
   let server: ApolloServer<ContextManager>;
   let url: string;
 
-  const UNARCHIVE_MUTATION = gql`
-    mutation saveUnArchive($id: [ID!]!, $timestamp: ISOString!) {
-      saveUnArchive(id: $id, timestamp: $timestamp) {
+  const SAVE_UNFAVORITE = gql`
+    mutation saveUnFavorite($id: [ID!]!, $timestamp: ISOString!) {
+      saveUnFavorite(id: $id, timestamp: $timestamp) {
         save {
           id
-          archived
-          archivedAt
+          favorite
+          favoritedAt
           updatedAt
         }
         errors {
@@ -40,20 +40,21 @@ describe('saveUnArchive mutation', function () {
   beforeEach(async () => {
     await writeDb('list').truncate();
     const inputData = [
-      { item_id: 0, status: 1, favorite: 0 },
-      { item_id: 1, status: 1, favorite: 0 },
-      { item_id: 2, status: 0, favorite: 1 },
+      { item_id: 0, favorite: 1 },
+      { item_id: 1, favorite: 1 },
+      { item_id: 2, favorite: 0 },
     ].map((row) => {
       return {
         ...row,
+        status: 0,
         user_id: 1,
         resolved_id: row.item_id,
         given_url: `http://${row.item_id}`,
         title: `title ${row.item_id}`,
         time_added: date,
         time_updated: date1,
-        time_read: row.status === 1 ? date : '0000-00-00 00:00:00',
-        time_favorited: date,
+        time_read: date,
+        time_favorited: row.favorite === 1 ? date : '0000-00-00 00:00:00',
         api_id: 'apiid',
         api_id_updated: 'apiid',
       };
@@ -74,7 +75,7 @@ describe('saveUnArchive mutation', function () {
 
   afterEach(() => jest.clearAllMocks());
 
-  it('should unarchive one save', async () => {
+  it('should unfavorite one save', async () => {
     const testTimestamp = '2023-10-05T14:48:00.000Z';
     const variables = {
       id: ['1'],
@@ -84,45 +85,21 @@ describe('saveUnArchive mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(UNARCHIVE_MUTATION), variables });
+      .send({ query: print(SAVE_UNFAVORITE), variables });
 
     expect(res).not.toBeUndefined();
-    expect(res.body.data.saveUnArchive.save).toBeArrayOfSize(1);
-    expect(res.body.data.saveUnArchive.errors).toBeArrayOfSize(0);
-    const actual = res.body.data.saveUnArchive.save[0];
+    expect(res.body.data.saveUnFavorite.save).toBeArrayOfSize(1);
+    expect(res.body.data.saveUnFavorite.errors).toBeArrayOfSize(0);
+    const actual = res.body.data.saveUnFavorite.save[0];
     expect(actual).toStrictEqual({
       id: '1',
-      archived: false,
-      archivedAt: null,
+      favorite: false,
+      favoritedAt: null,
       updatedAt: testTimestamp,
     });
   });
 
-  it('should fail the entire batch if one fails (NOT_FOUND)', async () => {
-    const testTimestamp = '2023-10-05T14:48:00.000Z';
-    const variables = {
-      id: ['123123'],
-      timestamp: testTimestamp,
-      updatedAt: testTimestamp,
-    };
-
-    const res = await request(app)
-      .post(url)
-      .set(headers)
-      .send({ query: print(UNARCHIVE_MUTATION), variables });
-
-    expect(res).not.toBeUndefined();
-    expect(res.body.data.saveUnArchive.save).toBeArrayOfSize(0);
-    const errors = res.body.data.saveUnArchive.errors;
-    expect(errors).toBeArrayOfSize(1);
-    expect(errors[0]).toStrictEqual({
-      __typename: 'NotFound',
-      message: 'Entity identified by key=id, value=123123 was not found.',
-      path: 'saveUnArchive',
-    });
-  });
-
-  it('should unarchive multiple saves', async () => {
+  it('should unfavorite multiple saves', async () => {
     const testTimestamp = '2023-10-05T14:48:00.000Z';
     const variables = {
       id: ['0', '1'],
@@ -132,31 +109,54 @@ describe('saveUnArchive mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(UNARCHIVE_MUTATION), variables });
+      .send({ query: print(SAVE_UNFAVORITE), variables });
 
     const expected = {
       save: [
         {
           id: '0',
-          archived: false,
-          archivedAt: null,
+          favorite: false,
+          favoritedAt: null,
           updatedAt: testTimestamp,
         },
         {
           id: '1',
-          archived: false,
-          archivedAt: null,
+          favorite: false,
+          favoritedAt: null,
           updatedAt: testTimestamp,
         },
       ],
       errors: [],
     };
-    const data = res.body.data.saveUnArchive;
+    const data = res.body.data.saveUnFavorite;
     expect(data.save).toIncludeSameMembers(expected.save);
     expect(data.errors).toBeArrayOfSize(0);
   });
 
-  it('should not fail if trying to unarchive a save that is already archived (no-op)', async () => {
+  it('should fail the entire batch if one fails (NOT_FOUND)', async () => {
+    const testTimestamp = '2023-10-05T14:48:00.000Z';
+    const variables = {
+      id: ['123123'],
+      timestamp: testTimestamp,
+    };
+
+    const res = await request(app)
+      .post(url)
+      .set(headers)
+      .send({ query: print(SAVE_UNFAVORITE), variables });
+
+    expect(res).not.toBeUndefined();
+    expect(res.body.data.saveUnFavorite.save).toBeArrayOfSize(0);
+    const errors = res.body.data.saveUnFavorite.errors;
+    expect(errors).toBeArrayOfSize(1);
+    expect(errors[0]).toStrictEqual({
+      __typename: 'NotFound',
+      message: 'Entity identified by key=id, value=123123 was not found.',
+      path: 'saveUnFavorite',
+    });
+  });
+
+  it('should not fail if trying to unfavorite a save that is already unfavorited (no-op)', async () => {
     const testTimestamp = '2023-10-05T14:48:00.000Z';
     const variables = {
       id: ['2'],
@@ -166,16 +166,17 @@ describe('saveUnArchive mutation', function () {
     const res = await request(app)
       .post(url)
       .set(headers)
-      .send({ query: print(UNARCHIVE_MUTATION), variables });
+      .send({ query: print(SAVE_UNFAVORITE), variables });
 
-    const data = res.body.data.saveUnArchive.save;
+    const data = res.body.data.saveUnFavorite.save;
     expect(data).toMatchObject([
       {
-        archived: false,
-        archivedAt: null,
-        updatedAt: date1.toISOString(),
+        favorite: false,
+        favoritedAt: null,
       },
     ]);
-    expect(res.body.data.saveUnArchive.errors).toBeArrayOfSize(0);
+    expect(res.body.data.saveUnFavorite.errors).toBeArrayOfSize(0);
   });
+  //todo: event emission
+  //todo: constraint validation
 });
