+++ servers/list-api/src/test/graphql/mutations/savedItemById/savedItemsMutationService-updates.integration.ts
@@ -1,7 +1,7 @@
-import { readClient, writeClient } from '../../../database/client';
-import { EventType } from '../../../businessEvents';
-import { ContextManager } from '../../../server/context';
-import { startServer } from '../../../server/apollo';
+import { readClient, writeClient } from '../../../../database/client';
+import { EventType } from '../../../../businessEvents';
+import { ContextManager } from '../../../../server/context';
+import { startServer } from '../../../../server/apollo';
 import { Application } from 'express';
 import { ApolloServer } from '@apollo/server';
 import request from 'supertest';
