+++ servers/list-api/src/test/graphql/mutations/tagsById/tagsMutationService-delete.integration.ts
@@ -1,8 +1,8 @@
-import { readClient, writeClient } from '../../../database/client';
-import { UsersMetaService } from '../../../dataService';
-import { EventType } from '../../../businessEvents';
-import { ContextManager } from '../../../server/context';
-import { startServer } from '../../../server/apollo';
+import { readClient, writeClient } from '../../../../database/client';
+import { UsersMetaService } from '../../../../dataService';
+import { EventType } from '../../../../businessEvents';
+import { ContextManager } from '../../../../server/context';
+import { startServer } from '../../../../server/apollo';
 import { Application } from 'express';
 import { ApolloServer } from '@apollo/server';
 import request from 'supertest';
